feat(category) add `ASetFeaturesCommandSpecificFigure`
The following is the rational for the prepend of `A`:

The `EnrichedFigure` inherited classes will be used in
a alphanumerical sorting, and this ordering is dictating the matching of figures
with classes.

See method `FromFigureDocument.get_figure_enriching_classes` for the
actual implementation of the list.

Since the functionality of matching figures with classes is by lazy regex
match(Only search for first match, break out and use it), the name of the
classes will have an impact on the ordering and thus the functionality.

For ease of implementation some class names will be prepended with a capitol `A`
to fix its ordering for matching.

Signed-off-by: Morten B. Rasmussen <m.rasmussen@samsung.com>